requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
pytz
python-dateutil
